fix(docs): new books dto file and minor fixes
