+++ src/book/dto/book.dto.ts
@@ -163,6 +163,7 @@ export class GetBooksDto extends BookFilterDto {
 export class GetBookByISBNDto {
   @IsNotEmpty()
   @IsISBN()
+  @ApiProperty({ example: '9789877978889' })
   isbn: string
 }
 
@@ -197,47 +198,4 @@ export class UpdateBookDto {
   @IsOptional()
   @IsArray()
   categories?: string[] | null
-}
-
-export class BooksKpi {
-  @IsNumber()
-  @ApiProperty()
-  totalBooks: number
-
-  @IsNumber()
-  @ApiProperty()
-  totalBooksMarkedAsDeleted: number
-
-  @IsNumber()
-  @ApiProperty()
-  totalPrices: number
-
-  @IsNumber()
-  @ApiProperty()
-  totalPricesMarkedAsDeleted: number
-
-  @IsNumber()
-  @ApiProperty()
-  totalAuthors: number
-
-  @IsNumber()
-  @ApiProperty()
-  totalCategories: number
-}
-
-export class BooksResponse {
-  @ApiProperty()
-  total: number
-
-  @ApiProperty()
-  page: number
-
-  @ApiProperty()
-  limit: number
-
-  @ApiProperty()
-  hasNextPage: boolean
-
-  @ApiProperty({ type: [BookDto] })
-  data: BookDto[]
 }
