+++ src/book/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './book.dto'
+export * from './book.response.dto'
